Separate log building logic out of ReplayES
